gas_ai_system/src/requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
joblib
firebase-admin
requests
pydantic
flask
flask-cors
huggingface_hub
python-dotenv
tqdm